transpile commutable and keep flow types
